rename: style.ts 파일명 수정 #62
