+++ frontend/src/components/Monitoring/InfoSection/InfoBox.tsx
@@ -1,4 +1,4 @@
-import * as S from './style';
+import * as S from './InfoSection.style';
 import cctvIcon from '@/assets/icons/cctvIcon.svg';
 import cctvIconBlue from '@/assets/icons/cctvBlueIcon.svg';
