+++ frontend/src/components/Monitoring/InfoSection/InfoSection.style.ts
